ON-15198: Revert `CI_EFCT_MAX_SUPERBUFS` to 512
Userspace applications compiled with older headers will fail due
to a mismatch in the value of `CI_EFCT_MAX_SUPERBUFS` between user
and kernel space. Therefore we revert its value to 512 and any code
affected by this change.

It is worth noting that applications relying on this patch will still
need to use the changes onload in userspace to avoid the userspace free
queue overfillng. That said, they should still be able to run; though an
assertion will be hit if the freeq is not serviced quickly enough by the
kernel (which only happens on a `napi_poll`). `napi_poll` is infrequent
when there are no incoming packets on the nic and so in this case the
freeq is likely to overflow on a large catchup.